updated readme and tidy comments
