--- helper.go
@@ -1,37 +1,0 @@
-package main
-
-import (
-	"net/http"
-
-	uuid "github.com/google/uuid"
-)
-
-func getUser(res http.ResponseWriter, req *http.Request) user {
-	// get current session cookie
-	myCookie, err := req.Cookie("myCookie")
-	if err != nil {
-		id := uuid.New()
-		myCookie = &http.Cookie{
-			Name:  "myCookie",
-			Value: id.String(),
-		}
-	}
-	http.SetCookie(res, myCookie)
-
-	// if the user exists already, get user
-	var myUser user
-	if username, ok := mapSessions[myCookie.Value]; ok {
-		myUser = mapUsers[username]
-	}
-	return myUser
-}
-
-func alreadyLoggedIn(req *http.Request) bool {
-	myCookie, err := req.Cookie("myCookie")
-	if err != nil {
-		return false
-	}
-	username := mapSessions[myCookie.Value]
-	_, ok := mapUsers[username]
-	return ok
-}
